修好了异常的brk以及随之导致的mmap,usertrap问题.问题是exec的时候栈的内容填错了

+++ kernel/include/stat.h
@@ -22,7 +22,7 @@ struct kstat {
         uint32 st_gid;
         uint64 st_rdev;
         unsigned long __pad;
-        unsigned int st_size;
+        long int st_size;
         uint32 st_blksize;
         int __pad2;
         uint64 st_blocks;
